pyproject.toml
[tool.poetry]
name = "parrot"
version = "0.1.0"
description = ""
authors = ["garlicOS® <14845347+Grosserly@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = {extras = ["voice"], version = "^1.5.1"}
python-dotenv = "^0.15.0"
jishaku = "^1.19.2"
markovify = "^0.8.3"
requests = "^2.25.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pyflakes = "^2.2.0"
flake8 = "^3.8.4"
rope = "^0.18.0"
mypy = "^0.790"
"discord.py-stubs" = "^1.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
